+++ src/paint_server.rs
@@ -7,7 +7,7 @@ use crate::render::prelude::*;
 pub fn fill(
     tree: &usvg::Tree,
     fill: &usvg::Fill,
-    bbox: Rect,
+    bbox: PathBbox,
     path: &tiny_skia::Path,
     anti_alias: bool,
     blend_mode: tiny_skia::BlendMode,
@@ -61,7 +61,7 @@ pub fn fill(
 pub fn stroke(
     tree: &usvg::Tree,
     stroke: &Option<usvg::Stroke>,
-    bbox: Rect,
+    bbox: PathBbox,
     path: &tiny_skia::Path,
     anti_alias: bool,
     blend_mode: tiny_skia::BlendMode,
@@ -134,7 +134,7 @@ pub fn stroke(
 fn prepare_linear(
     g: &usvg::LinearGradient,
     opacity: usvg::Opacity,
-    bbox: Rect,
+    bbox: PathBbox,
     paint: &mut tiny_skia::Paint,
 ) {
     let mode = match g.spread_method {
@@ -145,6 +145,9 @@ fn prepare_linear(
 
     let transform = {
         if g.units == usvg::Units::ObjectBoundingBox {
+            let bbox = try_opt_warn_or!(bbox.to_rect(), (),
+                "Gradient on zero-sized shapes is not allowed.");
+
             let mut ts = usvg::Transform::from_bbox(bbox);
             ts.append(&g.transform);
             ts.to_native()
@@ -177,7 +180,7 @@ fn prepare_linear(
 fn prepare_radial(
     g: &usvg::RadialGradient,
     opacity: usvg::Opacity,
-    bbox: Rect,
+    bbox: PathBbox,
     paint: &mut tiny_skia::Paint,
 ) {
     let mode = match g.spread_method {
@@ -188,6 +191,9 @@ fn prepare_radial(
 
     let transform = {
         if g.units == usvg::Units::ObjectBoundingBox {
+            let bbox = try_opt_warn_or!(bbox.to_rect(), (),
+                "Gradient on zero-sized shapes is not allowed.");
+
             let mut ts = usvg::Transform::from_bbox(bbox);
             ts.append(&g.transform);
             ts.to_native()
@@ -223,9 +229,12 @@ fn prepare_pattern_pixmap(
     pattern_node: &usvg::Node,
     pattern: &usvg::Pattern,
     global_ts: &usvg::Transform,
-    bbox: Rect,
+    bbox: PathBbox,
 ) -> Option<(tiny_skia::Pixmap, usvg::Transform)> {
     let r = if pattern.units == usvg::Units::ObjectBoundingBox {
+        let bbox = try_opt_warn_or!(bbox.to_rect(), None,
+            "Pattern on zero-sized shapes is not allowed.");
+
         pattern.rect.bbox_transform(bbox)
     } else {
         pattern.rect