+++ src/path.rs
@@ -9,7 +9,7 @@ pub fn draw(
     path: &usvg::Path,
     blend_mode: tiny_skia::BlendMode,
     canvas: &mut Canvas,
-) -> Option<Rect> {
+) -> Option<PathBbox> {
     let bbox = path.data.bbox();
     if path.visibility != usvg::Visibility::Visible {
         return bbox;
@@ -18,7 +18,7 @@ pub fn draw(
     // `usvg` guaranties that path without a bbox will not use
     // a paint server with ObjectBoundingBox,
     // so we can pass whatever rect we want, because it will not be used anyway.
-    let style_bbox = bbox.unwrap_or_else(|| Rect::new(0.0, 0.0, 1.0, 1.0).unwrap());
+    let style_bbox = bbox.unwrap_or_else(|| PathBbox::new(0.0, 0.0, 1.0, 1.0).unwrap());
 
     let skia_path = convert_path(&path.data)?;